# Add threads to the `ls` command in order to increase performance in some circumstances