@NotTheDr01ds 2024-09-13
That's a pretty big deal.  Could `glob` be optimized the same way?  As you said, WSL performance is abysmal.  I can't believe WSL2 has been out 4 years now and they still haven't found a way to fix that.