@fdncred 2024-09-13
We'd have to vendor the wax crate to make it multithreaded. I have an issue open about that over there I think, iirc.