@fdncred 2024-09-24
I'm a little bit on-the-fence about this one but let's land it and dogfood it for a while to see how it turns out. We can always revert if we need to.